pyproject.toml
[project]
name = "haru-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12, <3.14"
dependencies = [
    "anthropic>=0.83.0",
    "caldav>=1.3.0",
    "notion-client>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0",
    "vobject>=0.9.6",
]
